refactor: MermaidPreviewControls component
